+++ backend-project/small_eod/dictionaries/tests.py
@@ -1,28 +1,9 @@
 from django.test import TestCase
-
-from .factories import DictionaryFactory, FeatureFactory
-from .models import Dictionary, Feature
-from .serializers import DictionarySerializer
-from ..generic.tests import FactoryCreateObjectsMixin
-from ..users.factories import UserFactory
 from rest_framework.test import APIRequestFactory, force_authenticate
 
-
-class DictionaryFactoryTestCase(FactoryCreateObjectsMixin, TestCase):
-    MODEL = Dictionary
-    FACTORY = DictionaryFactory
-
-
-class FeatureFactoryTestCase(FactoryCreateObjectsMixin, TestCase):
-    MODEL = Feature
-    FACTORY = FeatureFactory
-
-    def test_foreign_keys(self):
-        """
-        Foreign keys are created.
-        """
-        obj = self.create_factory()
-        self.assertIsNotNone(obj.dictionary)
+from .models import Feature
+from .serializers import DictionarySerializer
+from ..users.factories import UserFactory
 
 
 class DictionarySerializerTestCase(TestCase):
